requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
yfinance==0.2.32
pandas==2.1.1
plotly==5.17.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0